Extract Steps component
