+++ src/pages/Registration/Dashboard/Steps/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Steps";
